% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{add_changepoints_scaled}
\alias{add_changepoints_scaled}
\title{Add scaled jump}
\usage{
add_changepoints_scaled(scenario_number, gamma, signal_to_noise, x_matrix, s)
}
\arguments{
\item{scenario_number}{1 : \eqn{\gamma \times p/2} dimensions undergo jump at t=.25, separate  \eqn{\gamma \times p/2} jump at t=0.75
2: Evenly spaced jumps between t=0.2 t=-.8 for \eqn{\gamma} dimensions undergoing jumps
10: first dimension undergoes a single jump at t=0.5}

\item{gamma}{proportion of dimensions undergoing a jump}

\item{signal_to_noise}{size of jump relative to local sd}

\item{x_matrix}{\eqn{p\times n} matrix of timme series}

\item{s}{vector of scales}
}
\value{
list, first entry matrix which can be added to x_matrix to add jumps, second matrix is matrix with two columns 'dim', 't'
tracking where jumps occur.
}
\description{
For the simulation scenarios used, add the scaled jumps to the time series
}
\examples{
x_matrix = ajdn::generate_x_matrix(1000,100,'IID',c(.05,.06))
add_cp = ajdn::add_changepoints_scaled(1,0.5,1,x_matrix,c(.05,.06))
}
