% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{bootstrap_lrv}
\alias{bootstrap_lrv}
\title{Multiplier bootstrap long-run variance}
\usage{
bootstrap_lrv(s_prime_list, autocovariances)
}
\arguments{
\item{s_prime_list}{vector of \eqn{s'} (block bootstrap sizes) that we are interested in}

\item{autocovariances}{vector of autocovariances, first entry is variance, will need a maximum autocovariance of ns'-1}
}
\value{
a vector of long run variances of multiplier bootstrap corresponding to the different \eqn{s'} we tested
}
\description{
This function estimates the long run variance of the multiplier bootstrap
}
\examples{
s_prime_list = c(1,2,3,4,5)
autocovariances = c(1,1,.5,.25,-.25,0,0,0,0,0,0,0)
ajdn::bootstrap_lrv(s_prime_list,autocovariances)
}
