% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{calc_bic}
\alias{calc_bic}
\title{Calculate BIC}
\usage{
calc_bic(time_series, cps_detected, s_prime)
}
\arguments{
\item{time_series}{a vector which is a time series}

\item{cps_detected}{dataframe of detected jumps (with column t_index)}

\item{s_prime}{\eqn{s'}}
}
\value{
numerical calcluation for BIC
}
\description{
This function calculates BIC on an individual dimension of a time series based on the changepoints that were detected
}
