% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{calc_desired_lrv_ratio}
\alias{calc_desired_lrv_ratio}
\title{Calculate desired LRV ratio}
\usage{
calc_desired_lrv_ratio(n)
}
\arguments{
\item{n}{length of time series}
}
\value{
LRV ratio (a scalar)
}
\description{
This funtion calculates the desired LRV ratio based on n (see Appendix B.1 on selection of hyperparameters for further description).
}
\examples{
ajdn::calc_desired_lrv_ratio(2000)
}
