% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{calc_local_sd}
\alias{calc_local_sd}
\title{Calculate local standard deviations}
\usage{
calc_local_sd(noise_matrix, s)
}
\arguments{
\item{noise_matrix}{matrix of random noise being used}

\item{s}{either vector of scales (if same for all dimension), or list of scales if different across dimensions used for testing}
}
\value{
\eqn{p \times n} matrix (same size as noise matrix) where each entry estimate of local standard deviation at that time
}
\description{
This function calculates the local standard deviation matrix for a time series. This
}
\examples{
noise_matrix = ajdn::generate_x_matrix(1000,100,'IID',c(.05,.06))
local_sd = ajdn::calc_local_sd(noise_matrix,c(.05,.06))
}
