% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJDN_Functions.R
\name{create_W_filter}
\alias{create_W_filter}
\title{Create W filter}
\usage{
create_W_filter(n, scale_count, t_index)
}
\arguments{
\item{n}{length of full time series}

\item{scale_count}{rounded value of n*s where s is the scale}

\item{t_index}{the index associated with the time at the center of the filter}
}
\value{
a vector of weights
}
\description{
Creates a vector of the optimal jump pass filter for a specific scale and time.
}
\examples{
w = create_W_filter(100,10,50)
plot(w)
}
