% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquake_data}
\alias{earthquake_data}
\title{2023 Turkey-Syria Earthquake Data}
\format{
\subsection{\code{earthquake_data}}{

A data frame with 28 rows and 5760 columns:
\describe{
\item{row names}{The name of the sensor the observations are taken from}
\item{columns}{Data sampled once every 10 seconds.}
...
}
}
}
\source{
\url{http://orfeus-eu.org/webdc3/}
}
\usage{
earthquake_data
}
\description{
A subset of data from seismic sensors during the 2023 Turkey-Syria earthquake.
Data from the ORFEUS Data Center WebDC3 Web Interface http://orfeus-eu.org/webdc3/.
The data is centred and the log of the absolute value is taken.
}
\keyword{datasets}
