% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquake_meta}
\alias{earthquake_meta}
\title{2023 Turkey-Syria Earthquake Meta-Data}
\format{
\subsection{\code{earthquake_meta}}{

A data frame with 28 rows and 29 columns:
\describe{
\item{sensor}{Short form of the sensor name}
\item{maxgapseconds}{Maximum gap of downtime in seconds.}
\item{numtraces}{number of traces in seismic data (1 trace = 0 downtime)}
\item{starttime}{starttime of observations}
\item{endtime}{endtime of observations}
\item{filename}{name of file downloaded from ORFEUS Data Center WebDC3 Web Interfact}
\item{latitude}{latitude of the sensor}
\item{longitude}{longitude of the sensor}
\item{location}{location name where the sensor is located}
...
}
}
}
\source{
\url{http://orfeus-eu.org/webdc3/}
}
\usage{
earthquake_meta
}
\description{
Metadata on each of the sensors used to collect information during the 2023 Turkey-Syria earthquake.
}
\keyword{datasets}
