% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{evaluate_changepoints}
\alias{evaluate_changepoints}
\title{Evaluate detected jumps}
\usage{
evaluate_changepoints(
  estimated_changepoint_locations,
  true_changepoint_locations,
  margin_of_error,
  p,
  n
)
}
\arguments{
\item{estimated_changepoint_locations}{a dataframe with jump locations, this dataframe must have columns 'dim' and 't'
representing the dimension the jump occured in and the time, in  [0,1] that it occurred. Note it can contain other columns
but this will not affect the evaluation}

\item{true_changepoint_locations}{the locations of the jump (as they were added when the data was generated), contains column 'dim' and 't'}

\item{margin_of_error}{tolerance for how far away from true value, an estimated value can be for it to be considered a successful
idenification, units in time in [0,1] (e.g. could be .0001)}

\item{p}{dimension of times series}

\item{n}{length of time series}
}
\value{
a vector with a number of performance metrics
}
\description{
This functions takes a set of estimated jump locations, and a set of
true jump locations, and returns a number of performance metrics
}
