% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{evaluate_changepoints_compare}
\alias{evaluate_changepoints_compare}
\title{Evaluate detected jumps}
\usage{
evaluate_changepoints_compare(
  true_changepoint_locations,
  margin_of_error,
  results_df,
  changepoints_by_method_list,
  scenario,
  n,
  K,
  signal_to_noise,
  gamma
)
}
\arguments{
\item{true_changepoint_locations}{a df with the dim and time of the true jumps}

\item{margin_of_error}{a value within [0,1] about how far away a detected jump can be to be labelled as correct}

\item{results_df}{a dataframe with parameters from the simulation, and what results we are recording}

\item{changepoints_by_method_list}{a list which has the jumps detected by different methods the entries are as follows
1st entry INSPECT , 2nd method DBLCUSUM, 3rd entry LOCLIN, 4th entry AJDN. First 3 entries are vectors, 4th entry a dataframe
with dimension and location of jumps.}

\item{scenario}{power scenario}

\item{n}{length of time series}

\item{K}{number of experiments}

\item{signal_to_noise}{SNR of jump}

\item{gamma}{proportion of dimensions undergoing jump}
}
\value{
dataframe of results
}
\description{
This function takes a list of jumps identified by the four methods tested and compares them to the true jump
locations, and then updates results_df appropriately based on the scenario note signal_to_noise an input to deal with case where jump_size=0
}
