% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{generate_means}
\alias{generate_means}
\title{Generate time varying means}
\usage{
generate_means(noise_matrix, s)
}
\arguments{
\item{noise_matrix}{matrix of random noise being used}

\item{s}{either vector of scales (if same for all dimension), or list of scales if different across dimensions used for testing}
}
\value{
either a \eqn{p \times n} matrix of the time varying mean
}
\description{
This function generates sin curve means with slightly different phases in each dimension
}
\examples{
noise_matrix = ajdn::generate_x_matrix(1000,100,'IID',c(.05,.06))
mean_matrix = ajdn::generate_means(noise_matrix,c(.05,.06))
plot(mean_matrix[1,],main='Scaled mean added to dimension 1',xlab='',ylab='')
}
