% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{generate_x_matrix}
\alias{generate_x_matrix}
\title{Generate simulated data}
\usage{
generate_x_matrix(n, p, dgp, s, x_matrix_split = NULL)
}
\arguments{
\item{n}{length of time series}

\item{p}{dimension of time series}

\item{dgp}{data generating process, one of: IID, IIDT, GS, GST, PS, PST, LS, LST, PLS, PLST}

\item{s}{either vector of scales (if same for all dimension), or list of scales if different across dimensions used for testing}

\item{x_matrix_split}{OPTIONAL argument, if true return list with first element the nosie matrix, and second
argument the potentially time varying mean matrix}
}
\value{
either a \eqn{p \times n} matrix of the time series, or a list of the noise matrix and mean matrix
}
\description{
This function generates simulated time series as used in the paper.
}
\examples{
ex_ts = ajdn::generate_x_matrix(1000,100,'GST',c(.05,.06))
}
