% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{power_summary_statistics}
\alias{power_summary_statistics}
\title{Power summary statistics}
\usage{
power_summary_statistics(
  results_df,
  K,
  scenario,
  test_type,
  gamma,
  s,
  p,
  jump_size,
  sim_result_dir
)
}
\arguments{
\item{results_df}{dataframe of resutls of power simulations}

\item{K}{number of experiments}

\item{scenario}{scenario from simulation}

\item{test_type}{type of test (power, type 1)}

\item{gamma}{gamma used}

\item{s}{scales used}

\item{p}{dimension of time series}

\item{jump_size}{size of jump for time series}

\item{sim_result_dir}{where to output results}
}
\description{
This function takes the result from K power experiments, and outputs summary statistics to a CSV file
}
