% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{preprocess_s}
\alias{preprocess_s}
\title{Preprocesses scales}
\usage{
preprocess_s(s, p, n)
}
\arguments{
\item{s}{vector of scales (e.g. c(0.03,0.05,0.07)), or list of length p with vector of scales for each entry}

\item{p}{dimension of time series}

\item{n}{length of time series}
}
\value{
a list of lists, and one vector,
1st entry - scale counts (scales in terms of number of observation,
e.g. n=100 s=0.05, scale count=5),
2nd entry - t indexes (indexes where we test for jumps, \eqn{T_r} in paper)
NOTE - 1st and 2nd entry each contain p elements, corresponding to each dimension.
3rd entry - vector of unique scales across all dimensions. The reason we create this is to
}
\description{
Preprocesses scales into counts and \eqn{T_r} for each dimension
}
\examples{
a = ajdn::preprocess_s(c(.02,.04),2,1000)
print(a)
}
