% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{s_prime_max_rot}
\alias{s_prime_max_rot}
\title{\eqn{s'_\text{max}} rule of thumb}
\usage{
s_prime_max_rot(n)
}
\arguments{
\item{n}{length of time series}
}
\value{
scalar number
}
\description{
Calculates rule of thumb values for \eqn{s'_\text{max}}
}
\examples{
ajdn::s_prime_max_rot(1000)
}
