% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{scales_rot}
\alias{scales_rot}
\title{Scale rule of thumbs}
\usage{
scales_rot(n, p)
}
\arguments{
\item{n}{length of time series}

\item{p}{dimension of time series}
}
\value{
vector of length 2 \eqn{(\underline{s},\bar{s})}
}
\description{
Calculates rule of thumb values for \eqn{(\underline{s},\bar{s})}
}
\examples{
ajdn::scales_rot(1000,10)
}
