% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJDN_Functions.R
\name{second_stage_refine}
\alias{second_stage_refine}
\title{Second stage refinement of detected jumps}
\usage{
second_stage_refine(first_stage_changepoints, x_matrix, scale_count_list)
}
\arguments{
\item{first_stage_changepoints}{a list where each entry of the list represents a jump with dim (dimension),
t_index (time where jump found), scale_count (ns where s scale used to find jump)
e.g. first_stage_changepoints[[5]]['t_index'] is the time index the 5th jump identified was found at}

\item{x_matrix}{p x n (p = # dimensions, n = # obsrevations) matrix of time series}

\item{scale_count_list}{a list of length p, where the pth entry of the list is a vector of \eqn{ns_{r,1},...,ns_{r,\delta_n}} where
\eqn{\delta_n} is the number of scales being used in dimension r}
}
\value{
a list of same length as first_stage_changepoints with the refined estimate of the time index where the jump occurs
}
\description{
Conducts the second stage procedure to refine estimation of jumps
}
