% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{single_local_sd}
\alias{single_local_sd}
\title{Calculates local standard deviation at poitn in time}
\usage{
single_local_sd(x_matrix, s_low, s_high, interested_dim, interested_t)
}
\arguments{
\item{x_matrix}{\eqn{p \times n} matrix of time series}

\item{s_low}{low scale for dimension adding jump}

\item{s_high}{high scale for dimension jumps is being added to}

\item{interested_dim}{which dimension we want local_sd for}

\item{interested_t}{which time (in (0,1)) we want local standard deviation for}
}
\value{
local standard deviation
}
\description{
This function takes a matrix of time series and then calculates for a specific dimension and specific time
what the local standard deviation is the reason behind having this function is a quick way to calculate local
standard deviation for the purposes of adding changepoints of the correct scale to simulated data.
}
\examples{
x_matrix = ajdn::generate_x_matrix(1000,100,'IID',c(.05,.06))
sd = ajdn::single_local_sd(x_matrix,.05,.07,1,.5)
}
