% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_Functions.R
\name{sparsify_scales}
\alias{sparsify_scales}
\title{Sparsify scales}
\usage{
sparsify_scales(s_low, s_high, n, p, max_scales = NULL, epsilon = 0.51)
}
\arguments{
\item{s_low}{lower scale \eqn{\underline{s}_r}}

\item{s_high}{upper scale \eqn{\bar{s}_r}}

\item{n}{length fo time series}

\item{p}{dimension of time series}

\item{max_scales}{OPTIONAL argument, maximum number of scales to return}

\item{epsilon}{default value = 0.51, used in calculation of \eqn{\delta_n}}
}
\value{
vector of length \eqn{\delta_n} \eqn{s_{r,1},\dots,s_{r,\delta_n}}
}
\description{
Calculates a sparse sequence of scales \eqn{s_{r,1},\dots,s_{r,\delta_n}}
}
\examples{
ajdn::sparsify_scales(0.05,0.08,1000,10)
}
