% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks_data}
\alias{stocks_data}
\title{Stocks Data}
\format{
\subsection{\code{stocks_data}}{

A data frame with 5 rows and 503 columns:
\describe{
\item{row names}{ticker symbol of stock}
\item{columns}{each column represents the daily price of a stock}
...
}
}
}
\source{
\url{https://www.quantmod.com/whatsnext/}
}
\usage{
stocks_data
}
\description{
Stocks data from 2021 to 2022 for META, AMZN, AAPL, NFLX, GOOG. The stock prices have been divided
by the value of the QQQ ETF to give a time series of the relative performance of the individual stocks
relative to the index.
}
\keyword{datasets}
