% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJDN_Functions.R
\name{time_series_normalization}
\alias{time_series_normalization}
\title{Calculate estimate of local standard deviation of time series}
\usage{
time_series_normalization(
  x_matrix,
  t_index_list,
  scale_count_list,
  num_cores = NULL,
  loud = NULL
)
}
\arguments{
\item{x_matrix}{p x n (p = # dimensions, n = # obsrevations) matrix of time series}

\item{t_index_list}{t_index_list a list of time indexes (\eqn{n t_j}) we need to calculate local standard deviations for for each dimension}

\item{scale_count_list}{a list of length \eqn{p}, where the \eqn{p}th entry of the list is a vector of \eqn{ns_{r,1},...,ns_{r,\delta_n}} where
\eqn{\delta_n} is the number of scales being used in dimension \eqn{r}}

\item{num_cores}{OPTIONAL argument, number of cores to use for parallelization}

\item{loud}{OPTIONAL, if TRUE print message when dimensions are normalized}
}
\value{
a dataframe, with the local sd, time index, and dimension
}
\description{
This function calcualtes an estimate of local standard deviation at each time point
}
