% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Functions.R
\name{var_1_LRV}
\alias{var_1_LRV}
\title{Autocov of VAR(1) model}
\usage{
var_1_LRV(coef_matrix, noise_cov, max_lag)
}
\arguments{
\item{coef_matrix}{\eqn{p \times p} matrix of VAR coefficients}

\item{noise_cov}{covariance matrix of the noise}

\item{max_lag}{maximum lag to return of the autocovariance}
}
\value{
array of autocovariance matrices
}
\description{
This function calculates the autocovariance array of a VAR_1 process using a truncated sum.
}
\examples{
autocov_array = ajdn::var_1_LRV(matrix(c(0.25,0.1,0.25,0.1),nrow=2),diag(c(1,1)),20)
print(autocov_array[,,1])
}
