% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJDN_Functions.R
\name{create_W_filter}
\alias{create_W_filter}
\title{Create W filter}
\usage{
create_W_filter(n, scale_count, t_index)
}
\arguments{
\item{n}{length of full time series}

\item{scale_count}{rounded value of \eqn{ns} where \eqn{s} is the scale}

\item{t_index}{the index associated with the time at the center of the filter (\eqn{t_jn})}
}
\value{
a vector of length \eqn{n} with weights that the filter takes
}
\description{
Creates a vector of the optimal jump pass filter for a \eqn{n} (length of time series), \eqn{s} (scale), and \eqn{t} (time)
}
\examples{
w = ajdn::create_W_filter(1000,100,300)
plot(w)
}
